Redo and submit completed landing page project
Both the index and style files were redone as there was no access
to the github repository while traveling.

Added containers in the header and testimonial section to improve
aesthetics and maintain spacing similar to sample photo provided.

Sections of the landing page were made to wrap to allow for better
viewing on smaller devices.